+++ validations_krt.go
@@ -1,0 +1,45 @@
+package main
+
+import "github.com/konstellation-io/krt/errors"
+
+func (krt *Krt) Validate() error {
+	var totalError error
+
+	err := krt.validateName()
+	totalError = errors.MergeErrors(totalError, err)
+
+	err = krt.validateDescription()
+	totalError = errors.MergeErrors(totalError, err)
+
+	err = krt.validateVersion()
+	totalError = errors.MergeErrors(totalError, err)
+
+	if len(krt.Workflows) == 0 {
+		totalError = errors.MergeErrors(
+			totalError,
+			errors.MissingRequiredFieldError("krt.workflows"),
+		)
+	} else {
+		for idx, workflow := range krt.Workflows {
+			err := workflow.Validate(idx)
+			totalError = errors.MergeErrors(totalError, err)
+		}
+	}
+
+	return totalError
+}
+
+func (krt *Krt) validateName() error {
+	return validateName(krt.Name, "krt.name")
+}
+
+func (krt *Krt) validateDescription() error {
+	if krt.Description == "" {
+		return errors.MissingRequiredFieldError("krt.description")
+	}
+	return nil
+}
+
+func (krt *Krt) validateVersion() error {
+	return validateName(krt.Version, "krt.version")
+}
